add api
création fichier routes.py
